+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -9,6 +9,9 @@
 ## 1.14.0.0
 
 * Add `mkTermToEvaluate` to `PlutusLanguage` class.
+* Add a field to `DebugFailure`
+* Convert `debugPlutus` to an `IO` action
+* Add `plutus-debug` executable
 
 ## 1.13.2.0